Autodownload Register
Nearly all the autodownload code was made by NarrowTux (Moritz
Schmale). All credits to him.
But I have found that, while autodownload works, errors happened while
trying to enable the plugin directly from the code. The only solution I
have found for it is to shutdown the server, telling the user to reopen
it. I expect someone will find a better way to do it.

The code that did the errors can be checked here:
https://gist.github.com/1307028