Update PermissionsBukkitHandler.java
PermissionsBukkit stores its data using UUIDs now, so you need to supply a UUID when attempting to retrieve data from it. However, to try and keep compatibility with older versions, I added a test to see if the current PermissionsBukkit being used is version 1.2, 1.5, 1.6, 2.0, 2.1, or 2.2 (these were the only old versions available on the devBukkit page for it). If it was one of the old versions, Essentials should supply a player name. If it's a newer version (such as the current PermissionsBukkit v2.3), then it will supply a UUID.

Note: Feel free to alter the way it checks for the PermissionsBukkit version or just scrap that altogether and make Essentials always supply a UUID~ I figured using a short RegEx string for compatibility would be helpful, so that's the only reason I added it.