Add support for colored mails.
I change it so that, if the user doesn't have permission to send colored mail, the mail will simply be stripped of format, as it was before. However, the major change here is that, if they have permission, it will replace the Essentials & color codes with the Minecraft § color code. The mail will then be stored in the recipient's .yml file (as previously), but with the § color codes. When the mail is printed, Minecraft will automatically replace the § color code with the corresponding color.